+++ scripts/upgradeSlave.sh
@@ -103,6 +103,9 @@ if [ "$MY_VERSION" == "1.0.1-stable" ]; then
   cp nafsdm/systemconfigs/nafscli /usr/bin/nafscli
   chmod +x /usr/bin/nafscli
 
+  # add new development section to config file (versions after 1.2.5)
+  echo -e "\n[development]\ngithub_branch = master\nskipVersionCheck = False\nincrementalCommitVersions = False" >> /home/slave-nafsdm/config.conf
+
   echo "* Upgrade completed."
   echo "* Script has automatically modified your config to match with the new standards."
 
@@ -127,6 +130,9 @@ elif [ "$MY_VERSION" == "1.1-stable" ]; then
   cp nafsdm/slave-nafsdm/nafscli /home/slave-nafsdm/nafscli -R
   cp nafsdm/systemconfigs/nafscli /usr/bin/nafscli
   chmod +x /usr/bin/nafscli
+
+  # add new development section to config file (versions after 1.2.5)
+  echo -e "\n[development]\ngithub_branch = master\nskipVersionCheck = False\nincrementalCommitVersions = False" >> /home/slave-nafsdm/config.conf
 
   echo "* Upgrade completed. You can now start nafsdm-slave again (make sure master is also upgraded!)-"
   rm -rf /home/slave-nafsdm/pythondaemon/tempUpgrade/temp_upgrade.sh
@@ -153,6 +159,9 @@ elif [ "$MY_VERSION" == "1.2-stable" ]; then
   cp nafsdm/systemconfigs/nafscli /usr/bin/nafscli
   chmod +x /usr/bin/nafscli
 
+  # add new development section to config file (versions after 1.2.5)
+  echo -e "\n[development]\ngithub_branch = master\nskipVersionCheck = False\nincrementalCommitVersions = False" >> /home/slave-nafsdm/config.conf
+
   echo "* Upgrade completed. You can now start nafsdm-slave again (make sure master is also upgraded!)-"
   rm -rf /home/slave-nafsdm/pythondaemon/tempUpgrade/temp_upgrade.sh
 
@@ -173,6 +182,9 @@ elif [ "$MY_VERSION" == "1.2.1-stable" ]; then
   cp nafsdm/systemconfigs/nafscli /usr/bin/nafscli
   chmod +x /usr/bin/nafscli
 
+  # add new development section to config file (versions after 1.2.5)
+  echo -e "\n[development]\ngithub_branch = master\nskipVersionCheck = False\nincrementalCommitVersions = False" >> /home/slave-nafsdm/config.conf
+
   echo "* Upgrade completed. You can now start nafsdm-slave again (make sure master is also upgraded!)-"
   rm -rf /home/slave-nafsdm/pythondaemon/tempUpgrade/temp_upgrade.sh
 
@@ -187,6 +199,9 @@ elif [ "$MY_VERSION" == "1.2.2-stable" ]; then
   cp nafsdm/slave-nafsdm/nafscli /home/slave-nafsdm/nafscli -R
   cp nafsdm/systemconfigs/nafscli /usr/bin/nafscli
   chmod +x /usr/bin/nafscli
+
+  # add new development section to config file (versions after 1.2.5)
+  echo -e "\n[development]\ngithub_branch = master\nskipVersionCheck = False\nincrementalCommitVersions = False" >> /home/slave-nafsdm/config.conf
 
   echo "* Upgrade completed. You can now start nafsdm-slave again (make sure master is also upgraded!)-"
   rm -rf /home/slave-nafsdm/pythondaemon/tempUpgrade/temp_upgrade.sh
@@ -203,6 +218,9 @@ elif [ "$MY_VERSION" == "1.2.3-stable" ]; then
   cp nafsdm/systemconfigs/nafscli /usr/bin/nafscli
   chmod +x /usr/bin/nafscli
 
+  # add new development section to config file (versions after 1.2.5)
+  echo -e "\n[development]\ngithub_branch = master\nskipVersionCheck = False\nincrementalCommitVersions = False" >> /home/slave-nafsdm/config.conf
+
   echo "* Upgrade completed. You can now start nafsdm-slave again (make sure master is also upgraded!)-"
   rm -rf /home/slave-nafsdm/pythondaemon/tempUpgrade/temp_upgrade.sh
 
@@ -218,6 +236,9 @@ elif [ "$MY_VERSION" == "1.2.4-stable" ]; then
   cp nafsdm/systemconfigs/nafscli /usr/bin/nafscli
   chmod +x /usr/bin/nafscli
 
+  # add new development section to config file (versions after 1.2.5)
+  echo -e "\n[development]\ngithub_branch = master\nskipVersionCheck = False\nincrementalCommitVersions = False" >> /home/slave-nafsdm/config.conf
+
   echo "* Upgrade completed. You can now start nafsdm-slave again (make sure master is also upgraded!)-"
   rm -rf /home/slave-nafsdm/pythondaemon/tempUpgrade/temp_upgrade.sh
 
@@ -231,6 +252,9 @@ elif [ "$MY_VERSION" == "1.2.5-stable" ]; then
   # nafscli reinstall (delete and copy) (from version 1.2.5 onwards)
   rm -rf /home/slave-nafsdm/nafscli
   cp nafsdm/slave-nafsdm/nafscli /home/slave-nafsdm/nafscli -R
+
+  # add new development section to config file (versions after 1.2.5)
+  echo -e "\n[development]\ngithub_branch = master\nskipVersionCheck = False\nincrementalCommitVersions = False" >> /home/slave-nafsdm/config.conf
 
   echo "* Upgrade completed. You can now start nafsdm-slave again (make sure master is also upgraded!)-"
   rm -rf /home/slave-nafsdm/pythondaemon/tempUpgrade/temp_upgrade.sh
@@ -254,13 +278,12 @@ elif [ "$MY_VERSION" == "dev_release" ]; then
     cd /tmp/nafsdm
     COMMIT_HASH=$(git log -n 1 development | sed -n '1p' | cut -c8-14)
     echo "version = \"$COMMIT_HASH-dev\"" > /home/slave-nafsdm/pythondaemon/version.py
-    echo "True" > /home/slave-nafsdm/pythondaemon/dev_ic_mode.txt
-    echo "development" > /home/slave-nafsdm/pythondaemon/dev_github_branch.txt
 
     echo "* Done."
   fi
 
   echo "* Upgrade completed. You can now start nafsdm-slave again (make sure master is also upgraded!)-"
+  echo "* Make sure to copy the [development] section into config.conf and make sure branch is set to development and that incrementalCommitVersions is set to True."
   rm -rf /home/slave-nafsdm/pythondaemon/tempUpgrade/temp_upgrade.sh
 
   exit 0